// client/models/AuditLogResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TechnicianResponse } from './TechnicianResponse';
export type AuditLogResponse = {
    id: string;
    technician: TechnicianResponse;
    action: string;
    timeStamp: string;
};