// client/models/AuditLogManyResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AuditLogResponse } from './AuditLogResponse';
export type AuditLogManyResponse = {
    data: Array<AuditLogResponse>;
    skip: number;
    take: number;
    rows: number;
};